requirements.txt
# Dependencias generales del proyecto
# requirements-full.txt
# Primero debes instalar Airflow manualmente (ver README)
#pip install "apache-airflow[gcp]==2.7.3" `
  #--constraint "https://raw.githubusercontent.com/apache/airflow/constraints-2.7.3/constraints-3.10.txt"
# Luego el resto de paquetes

google-cloud-bigquery==3.13.0
google-cloud-storage==2.10.0
google-cloud-pubsub==2.18.4
pandas==2.1.3
numpy==1.24.3
faker==20.1.0
kafka-python==2.0.2
pyspark==3.5.0
python-dotenv==1.0.0
pytest==7.4.3
pyyaml==6.0.1
sqlalchemy==2.0.23
